chore: indexer modularization
