+++ apps/indexer/src/lib/viem.ts
@@ -1,0 +1,1 @@
+export {}
