+++ apps/indexer/src/services/easyauction.ts
@@ -1,10 +1,10 @@
 import { getContract } from 'viem'
-import { client } from './viem-client'
+import { client } from '../viem'
 import { TestnetEasyAuction } from 'smartsale-contracts'
 
 export const easyAuction = getContract({
   ...TestnetEasyAuction,
-  client: { 
-    public: client, 
-  } 
+  client: {
+    public: client,
+  },
 })
